Updated integration for jcodec 2.3

+++ integration/boofcv-jcodec/src/main/java/boofcv/io/jcodec/ImplConvertJCodecPicture.java
@@ -39,7 +39,7 @@ public class ImplConvertJCodecPicture {
 
 		output.reshape(input.getWidth(),input.getHeight());
 
-		int dataIn[] = input.getData()[0];
+		byte dataIn[] = input.getData()[0];
 
 		GrayU8 out0 = output.getBand(0);
 		GrayU8 out1 = output.getBand(1);
@@ -63,9 +63,9 @@ public class ImplConvertJCodecPicture {
 
 	public static void yuv420_to_PlRgb_U8(Picture input, Planar<GrayU8> output) {
 
-		int[] Y = input.getPlaneData(0);
-		int[] U = input.getPlaneData(1);
-		int[] V = input.getPlaneData(2);
+		byte[] Y = input.getPlaneData(0);
+		byte[] U = input.getPlaneData(1);
+		byte[] V = input.getPlaneData(2);
 
 		GrayU8 R = output.getBand(0);
 		GrayU8 G = output.getBand(1);
@@ -105,9 +105,9 @@ public class ImplConvertJCodecPicture {
 
 	public static void yuv420_to_PlRgb_F32(Picture input, Planar<GrayF32> output) {
 
-		int[] Y = input.getPlaneData(0);
-		int[] U = input.getPlaneData(1);
-		int[] V = input.getPlaneData(2);
+		byte[] Y = input.getPlaneData(0);
+		byte[] U = input.getPlaneData(1);
+		byte[] V = input.getPlaneData(2);
 
 		GrayF32 R = output.getBand(0);
 		GrayF32 G = output.getBand(1);
@@ -147,7 +147,7 @@ public class ImplConvertJCodecPicture {
 
 	public static void yuv420_to_U8(Picture input, GrayU8 output) {
 
-		int[] Y = input.getPlaneData(0);
+		byte[] Y = input.getPlaneData(0);
 
 		final int yStride = output.width;
 
@@ -169,7 +169,7 @@ public class ImplConvertJCodecPicture {
 
 	public static void yuv420_to_F32(Picture input, GrayF32 output) {
 
-		int[] Y = input.getPlaneData(0);
+		byte[] Y = input.getPlaneData(0);
 
 		final int yStride = output.width;
